test: fold the repeated driver and dependency scaffolding into one fixture each
The CLI runner suite hand-wrote ten `MakaSessionDriver` implementations that
differed in one behaviour apiece and repeated the other fourteen members
verbatim, and the two `createAppShellChatActions` suites each carried a full
copy of the actions' wide dependency surface. Both shapes have to be edited
once per copy whenever the driver interface or the dependency list grows, and
a copy that is missed drifts silently rather than failing.

Give each its single owner: an abstract `FakeSessionDriver` base that supplies
the inert members and the Host-admission `submitMessage`, leaving subclasses
to state only the behaviour they exist to exercise; and a
`app-shell-chat-actions-fixture` module holding the window installer, the
turn and message state doubles, and the dependency factory.

No test behaviour changes; the CLI suite and both Desktop suites pass
unchanged.

Generated-by: Claude Code
