+++ apps/desktop/src/main/__tests__/app-shell-busy-race-settlement.test.ts
@@ -32,94 +32,13 @@ import type { StoredMessage } from '@maka/core/session';
 import type { LiveTurnProjection } from '@maka/ui';
 import { createAppShellChatActions } from '../../renderer/app-shell-chat-actions.js';
 
-function installWindow(maka: unknown): () => void {
-  const target = globalThis as unknown as { window?: unknown };
-  const hadWindow = Object.prototype.hasOwnProperty.call(target, 'window');
-  const previousWindow = target.window;
-  Object.defineProperty(target, 'window', {
-    configurable: true,
-    value: { maka },
-    writable: true,
-  });
-  return () => {
-    if (hadWindow) {
-      Object.defineProperty(target, 'window', {
-        configurable: true,
-        value: previousWindow,
-        writable: true,
-      });
-    } else {
-      delete target.window;
-    }
-  };
-}
-
-function createTurnState() {
-  const liveTurnBySession: Record<string, LiveTurnProjection> = {};
-  return {
-    liveTurnBySession,
-    setLiveTurnBySession(
-      updater: (c: Record<string, LiveTurnProjection>) => Record<string, LiveTurnProjection>,
-    ) {
-      const next = updater({ ...liveTurnBySession });
-      for (const key of Object.keys(liveTurnBySession)) delete liveTurnBySession[key];
-      Object.assign(liveTurnBySession, next);
-    },
-  };
-}
-
-function createMessageState() {
-  const messages: StoredMessage[] = [];
-  return {
-    messages,
-    setMessages(updater: StoredMessage[] | ((current: StoredMessage[]) => StoredMessage[])) {
-      const next = typeof updater === 'function' ? updater([...messages]) : updater;
-      messages.length = 0;
-      messages.push(...next);
-    },
-  };
-}
-
-function createActionsDeps() {
-  return {
-    uiLocale: 'en' as const,
-    activeIdRef: { current: undefined as string | undefined },
-    addPendingSessionAction: () => true,
-    captureComposerImportOwner: () => ({
-      sessionId: undefined,
-      navSection: 'sessions' as const,
-    }),
-    checkTaskSubmissionReadiness: async () => true,
-    clearPendingSessionAction: () => undefined,
-    isNewChatSendSurfaceActive: () => true,
-    isShellSurfaceOwnerActive: () => true,
-    markSessionReadLocally: () => undefined,
-    messageRetryPendingRef: { current: new Set<string>() },
-    refreshSessions: async () => [],
-    setActiveId: () => undefined,
-    setMessageLoadErrorBySession: () => undefined,
-    setMessageRetryPendingBySession: () => undefined,
-    setMessages: () => undefined,
-    addTransientMessage: () => undefined,
-    updateTransientMessage: () => undefined,
-    removeTransientMessage: () => undefined,
-    transcriptRangeRef: { current: undefined },
-    setNavSelection: () => undefined,
-    setLiveTurnBySession: () => undefined,
-    setInteractionBySession: () => undefined,
-    showModelSetupToast: () => undefined,
-    toastApi: { error: () => undefined, info: () => undefined },
-    newChatModel: null,
-    pendingNewChatThinkingLevel: null,
-    newChatPermissionChoice: undefined,
-    clearNewChatPermissionChoice: () => {},
-    newChatCollaborationMode: 'agent' as const,
-    newChatOrchestrationMode: 'default' as const,
-    newTaskTarget: { profileId: 'local', hostId: 'host-local', projectId: null },
-  };
-}
-
-const EMPTY_SKILL_INVOCATION = { loaded: [], failed: [], receipts: [] };
+import {
+  createActionsDeps,
+  createMessageState,
+  createTurnState,
+  EMPTY_SKILL_INVOCATION,
+  installWindow,
+} from './app-shell-chat-actions-fixture.js';
 
 describe('busy-raced send settlement', () => {
   it('shows a Follow Up immediately and keeps its caller-owned identity', async () => {
