+++ apps/desktop/src/main/__tests__/app-shell-chat-actions-fixture.ts
@@ -1,0 +1,123 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+/**
+ * Shared scaffolding for the `createAppShellChatActions` suites. The dependency
+ * surface is wide and the suites only ever vary a handful of entries, so a
+ * second copy of it drifts silently and has to be edited twice whenever the
+ * actions gain a dependency.
+ */
+
+import type { StoredMessage } from '@maka/core/session';
+import type { LiveTurnProjection } from '@maka/ui';
+
+/** Installs a `window.maka` bridge double; the returned function restores it. */
+export function installWindow(maka: unknown): () => void {
+  const target = globalThis as unknown as { window?: unknown };
+  const hadWindow = Object.prototype.hasOwnProperty.call(target, 'window');
+  const previousWindow = target.window;
+  Object.defineProperty(target, 'window', {
+    configurable: true,
+    value: { maka },
+    writable: true,
+  });
+  return () => {
+    if (hadWindow) {
+      Object.defineProperty(target, 'window', {
+        configurable: true,
+        value: previousWindow,
+        writable: true,
+      });
+    } else {
+      delete target.window;
+    }
+  };
+}
+
+/**
+ * The live-turn arm as a real map rather than a black-hole stub: a send that
+ * never lands must leave nothing behind, and that cannot be asserted against a
+ * no-op setter.
+ */
+export function createTurnState() {
+  const liveTurnBySession: Record<string, LiveTurnProjection> = {};
+  return {
+    liveTurnBySession,
+    setLiveTurnBySession(
+      updater: (c: Record<string, LiveTurnProjection>) => Record<string, LiveTurnProjection>,
+    ) {
+      const next = updater({ ...liveTurnBySession });
+      for (const key of Object.keys(liveTurnBySession)) delete liveTurnBySession[key];
+      Object.assign(liveTurnBySession, next);
+    },
+  };
+}
+
+export function createMessageState() {
+  const messages: StoredMessage[] = [];
+  return {
+    messages,
+    setMessages(updater: StoredMessage[] | ((current: StoredMessage[]) => StoredMessage[])) {
+      const next = typeof updater === 'function' ? updater([...messages]) : updater;
+      messages.length = 0;
+      messages.push(...next);
+    },
+  };
+}
+
+export function createActionsDeps() {
+  return {
+    uiLocale: 'en' as const,
+    activeIdRef: { current: undefined as string | undefined },
+    addPendingSessionAction: () => true,
+    captureComposerImportOwner: () => ({
+      sessionId: undefined,
+      navSection: 'sessions' as const,
+    }),
+    checkTaskSubmissionReadiness: async () => true,
+    clearPendingSessionAction: () => undefined,
+    isNewChatSendSurfaceActive: () => true,
+    isShellSurfaceOwnerActive: () => true,
+    markSessionReadLocally: () => undefined,
+    messageRetryPendingRef: { current: new Set<string>() },
+    refreshSessions: async () => [],
+    setActiveId: () => undefined,
+    setMessageLoadErrorBySession: () => undefined,
+    setMessageRetryPendingBySession: () => undefined,
+    setMessages: () => undefined,
+    addTransientMessage: () => undefined,
+    updateTransientMessage: () => undefined,
+    removeTransientMessage: () => undefined,
+    transcriptRangeRef: { current: undefined },
+    setNavSelection: () => undefined,
+    setLiveTurnBySession: () => undefined,
+    setInteractionBySession: () => undefined,
+    showModelSetupToast: () => undefined,
+    toastApi: { error: () => undefined, info: () => undefined },
+    newChatModel: null,
+    pendingNewChatThinkingLevel: null,
+    newChatPermissionChoice: undefined,
+    clearNewChatPermissionChoice: () => {},
+    newChatCollaborationMode: 'agent' as const,
+    newChatOrchestrationMode: 'default' as const,
+    newTaskTarget: { profileId: 'local', hostId: 'host-local', projectId: null },
+  };
+}
+
+export const EMPTY_SKILL_INVOCATION = { loaded: [], failed: [], receipts: [] };
